Remove time component
The time component was not interesting, instead we give the player
choice of a lot of items, with a sale timer. Every 5 seconds, some
amount of the player's items will sell. The player has the choice of
higher profit, or higher chance of sale.

This replaces the time component with a timed sale, but does not add any
of the following:
Automation
Upgrades
Prestige